--- app/__init__.py
@@ -1,16 +1,0 @@
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-from flask_migrate import Migrate
-from flask_login import LoginManager
-from flask_debugtoolbar import DebugToolbarExtension
-from config import Config
-
-app = Flask(__name__)
-app.config.from_object(Config)
-db = SQLAlchemy(app)
-migrate = Migrate(app, db)
-login = LoginManager(app)
-login.login_view = 'login'
-login.session_protection = 'strong'
-
-from app import routes, models
