--- requirements.txt
@@ -1,34 +1,0 @@
-alembic==1.0.0
-asn1crypto==0.24.0
-blinker==1.4
-certifi==2018.8.24
-cffi==1.11.5
-chardet==3.0.4
-click==6.7
-cryptography==2.3.1
-Flask==1.0.2
-Flask-DebugToolbar==0.10.1
-Flask-Login==0.4.1
-Flask-Migrate==2.2.1
-Flask-SQLAlchemy==2.3.2
-Flask-WTF==0.14.2
-idna==2.7
-itsdangerous==0.24
-Jinja2==2.10
-Mako==1.0.7
-MarkupSafe==1.0
-oauthlib==2.1.0
-pycparser==2.18
-PyMySQL==0.9.2
-pyOpenSSL==18.0.0
-python-dateutil==2.7.3
-python-dotenv==0.9.1
-python-editor==1.0.3
-rauth==0.7.3
-requests==2.19.1
-requests-oauthlib==1.0.0
-six==1.11.0
-SQLAlchemy==1.2.11
-urllib3==1.23
-Werkzeug==0.14.1
-WTForms==2.2.1
